feat(platform-api): extract package skeleton, move services factory out of platform-ui
Creates the `@mediforce/platform-api` package with subpath exports
(`/contract`, `/handlers`, `/services`, `/client`) and moves the
platform-services factory from `packages/platform-ui/src/lib/` into
`@mediforce/platform-api/services`. The UI keeps a thin shim at
`@/lib/platform-services` so ~100 existing route / action imports
continue to work unchanged until Phase 5 of the headless migration
codemods them.

Extracted helper `getAppBaseUrl` into its own `@/lib/app-base-url`
so the shim no longer mixes unrelated concerns.
