+++ packages/platform-api/src/contract/index.ts
@@ -1,0 +1,6 @@
+export {
+  ListTasksInputSchema,
+  ListTasksOutputSchema,
+  type ListTasksInput,
+  type ListTasksOutput,
+} from './tasks.js';
