+++ packages/platform-api/src/handlers/index.ts
@@ -1,0 +1,1 @@
+export { listTasks, type ListTasksDeps } from './tasks/list-tasks.js';
