+++ packages/platform-api/src/services/index.ts
@@ -1,0 +1,2 @@
+export { getPlatformServices, type PlatformServices } from './platform-services.js';
+export { seedBuiltinAgentDefinitions } from './seed-agent-definitions.js';
